added a sort function for an array of movies and a star for your favorite movies

+++ js/script.js
@@ -2,35 +2,57 @@
 
 document.addEventListener('DOMContentLoaded', () => {
 
-    const promoBG = document.querySelector('.promo__bg');
-    const banners = document.querySelectorAll('.promo__adv img');
-    const genre = promoBG.querySelector('.promo__genre');
-    const movieList = document.querySelector('.promo__interactive-list');
-    const addMovieForm = document.querySelector('.add');
-    const movieNameInput = addMovieForm.querySelector('.adding__input');
-    const favoriteMovieCheckbox = addMovieForm.querySelector(`input[type='checkbox']`);
+    const promoBG = document.querySelector('.promo__bg'),
+        banners = document.querySelectorAll('.promo__adv img'),
+        genre = promoBG.querySelector('.promo__genre'),
+        movieList = document.querySelector('.promo__interactive-list'),
+        addMovieForm = document.querySelector('.add'),
+        movieNameInput = addMovieForm.querySelector('.adding__input'),
+        favoriteMovieCheckbox = addMovieForm.querySelector(`input[type='checkbox']`);
+
+    // const movieDB = {
+    //     movies: [
+    //         "Логан",
+    //         "Лига справедливости",
+    //         "Ла-ла лэнд",
+    //         "Одержимость",
+    //         "Скотт Пилигрим против..."
+    //     ]
+    // };
 
     const movieDB = {
         movies: [
-            "Логан",
-            "Лига справедливости",
-            "Ла-ла лэнд",
-            "Одержимость",
-            "Скотт Пилигрим против..."
+            {name: "логан", favorite: false},
+            {name: "лига справедливости", favorite: false},
+            {name: "ла-ла лэнд", favorite: false},
+            {name: "одержимость", favorite: false},
+            {name: "скотт Пилигрим против...", favorite: false},
         ]
     };
 
-    promoBG.style.background = `url('img/bg.jpg')`;
-    genre.textContent = 'ДРАМА';
+    function MakeChanges() {
+        promoBG.style.background = `url('img/bg.jpg')`;
+        genre.textContent = 'ДРАМА';
+    }
 
-    function renderMovieList() {
-        movieList.innerHTML = '';
-        movieDB.movies.sort();
-        movieDB.movies.forEach((movie, index) => {
-            movieList.innerHTML += `<li class="promo__interactive-item">
-                                        ${++index}. ${movie}
-                                        <div class="delete"></div>
-                                    </li>`;
+    function renderMovieList(movies, movieContainer) {
+        movieContainer.innerHTML = '';
+        sortMovieArray(movies);
+        movies.forEach((movie, index) => {
+            
+            if(movie.favorite) {
+                console.log(movie.favorite);
+                movieContainer.innerHTML += `<li class="promo__interactive-item">
+                                                <span></span>
+                                                ${++index}. ${movie.name}
+                                                <div class="delete"></div>
+                                            </li>`;
+            } else {
+                movieContainer.innerHTML += `<li class="promo__interactive-item">
+                                                ${++index}. ${movie.name}
+                                                <div class="delete"></div>
+                                            </li>`;
+            }
         });
     }
 
@@ -40,11 +62,14 @@ document.addEventListener('DOMContentLoaded', () => {
 
     function addNewMovie(event) {
         event.preventDefault();
-        if(movieNameInput.value === '') return false;
-        isFavoriteMovie(favoriteMovieCheckbox);
-        movieDB.movies.push(checkLength(movieNameInput.value));
-        resetForm();
-        renderMovieList();
+        if(movieNameInput.value) {
+            let favorite = isFavoriteMovie(favoriteMovieCheckbox);
+            movieNameInput.value = checkLength(movieNameInput.value).toLowerCase();
+            movieDB.movies.push({name: movieNameInput.value, favorite: favorite});
+            renderMovieList(movieDB.movies, movieList);
+        }
+        event.target.reset();
+
     }
 
     function checkLength(str) {
@@ -52,19 +77,16 @@ document.addEventListener('DOMContentLoaded', () => {
     }
 
     function deleteMovieFromList(event) {
-        let listElement = event.target.closest('li');
-        movieDB.movies = deleteArrayElement(movieDB.movies, listElement.textContent.trim()[0] - 1);
-        listElement.remove();
-        renderMovieList();
-    }
-
-    function resetForm() {
-        movieNameInput.value = '';
-        favoriteMovieCheckbox.checked = false;
+        if(event.target.classList.contains('delete')) {
+            let listElement = event.target.closest('li');
+            movieDB.movies = deleteArrayElement(movieDB.movies, listElement.textContent.trim()[0] - 1);
+            listElement.remove();
+            renderMovieList(movieDB.movies, movieList);
+        }
     }
 
     function isFavoriteMovie(checkbox) {
-        if(checkbox.checked) console.log('Сделать любимым');
+        return checkbox.checked ? true : false;
     }
 
     function deleteArrayElement(array, elementIndex) {
@@ -73,9 +95,23 @@ document.addEventListener('DOMContentLoaded', () => {
         });
     }
 
+    function sortMovieArray(array) {
+        array.sort(function(a, b) {
+            if (a.name < b.name) {
+                return -1;
+              }
+              if (a.name > b.name) {
+                return 1;
+              }
+      
+              return 0;
+        });
+    }
+
     addMovieForm.addEventListener('submit', addNewMovie);
     movieList.addEventListener('click', deleteMovieFromList);
 
+    MakeChanges();
     removeElements(banners);
-    renderMovieList();
+    renderMovieList(movieDB.movies, movieList);
 });
